Refactor CDCResponseGenerator
